+++ Sources/Sheeeeeeeeet/ActionSheet/Presenters/ActionSheetPopoverPresenter+Deprecations.swift
@@ -1,0 +1,17 @@
+//
+//  ActionSheetPopoverPresenter+Deprecations.swift
+//  Sheeeeeeeeet
+//
+//  Created by Daniel Saidi on 2019-12-08.
+//
+
+import UIKit
+
+public extension ActionSheetPopoverPresenter {
+    
+    @available(*, deprecated, renamed: "isDismissableWithOrientationChange")
+    var isListeningToOrientationChanges: Bool {
+        get { isDismissableWithOrientationChange }
+        set { isDismissableWithOrientationChange = newValue }
+    }
+}