+++ Sources/Sheeeeeeeeet/ActionSheet/Presenters/ActionSheetPresenterBase.swift
@@ -1,0 +1,21 @@
+//
+//  ActionSheetPresenterBase.swift
+//  Sheeeeeeeeet
+//
+//  Created by Daniel Saidi on 2019-12-08.
+//  Copyright © 2019 Daniel Saidi. All rights reserved.
+//
+
+import UIKit
+
+/**
+ This is a base class for action sheet presenters. It's used
+ by the built-in presenters and can be used by yours as well.
+ */
+open class ActionSheetPresenterBase {
+
+    
+    // MARK: - Properties
+    
+    public var events = ActionSheetPresenterEvents()
+}